refactor: abstract out normalizeOptions
