+++ libs/tailwind/src/utils/add-config-files.ts
@@ -7,15 +7,16 @@ import {
   Rule,
   url,
 } from '@angular-devkit/schematics';
+import { NormalizedTailwindSchematicsOptions } from '../schematics/schema';
 import { isInJest } from './is-in-jest';
 
-export function addConfigFiles(
-  enableTailwindInComponentsStyles: boolean,
-  darkMode: 'none' | 'class' | 'media',
-  appsDir?: string,
-  libsDir?: string,
-  sourceRoot = 'src'
-): Rule {
+export function addConfigFiles({
+  enableTailwindInComponentsStyles,
+  darkMode,
+  appsDir,
+  libsDir,
+  sourceRoot = 'src',
+}: NormalizedTailwindSchematicsOptions): Rule {
   return mergeWith(
     apply(url(isInJest() ? '../files' : './files'), [
       applyTemplates({
